Removed attack vector users could configure themselves into.
"So if there was a DOS vector against them, it wouldn't be fixed and the nodes could be shut down remotely. After those are shut down, the ones that have it turned on explicitly might also die off because they are flooded with filters at that point. So effectively after 0.18 is EOL, this p2p feature should be considered dead."

https://github.com/bitcoin/bitcoin/pull/16152#issuecomment-511569067